fix can not find jest bug
