+++ .bitmap
@@ -1,7 +1,7 @@
 /* THIS IS A BIT-AUTO-GENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. */
 
 {
-    "bit.test-envs/internal/e2e-helper@0.0.18": {
+    "bit.test-envs/internal/e2e-helper@0.0.19": {
         "files": [
             {
                 "relativePath": "test/e2e-helper.ts",
@@ -13,7 +13,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/testers/jest@0.0.18": {
+    "bit.test-envs/testers/jest@0.0.19": {
         "files": [
             {
                 "relativePath": "src/jest/README.md",
@@ -110,7 +110,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/internal/test-util@0.0.18": {
+    "bit.test-envs/internal/test-util@0.0.19": {
         "files": [
             {
                 "relativePath": "test/envs-test-utils.ts",
@@ -122,7 +122,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/compilers/webpack@0.0.18": {
+    "bit.test-envs/compilers/webpack@0.0.19": {
         "files": [
             {
                 "relativePath": "src/webpack/README.md",
@@ -229,7 +229,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/internal/env-utils@0.0.18": {
+    "bit.test-envs/internal/env-utils@0.0.19": {
         "files": [
             {
                 "relativePath": "src/env-utils/compiler-utils.ts",
@@ -252,7 +252,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/testers/mocha@0.0.18": {
+    "bit.test-envs/testers/mocha@0.0.19": {
         "files": [
             {
                 "relativePath": "src/mocha/README.md",
@@ -329,7 +329,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/compilers/babel@0.0.18": {
+    "bit.test-envs/compilers/babel@0.0.19": {
         "files": [
             {
                 "relativePath": "src/babel/README.md",
@@ -436,7 +436,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/env-utils/find-configuration@0.0.18": {
+    "bit.test-envs/env-utils/find-configuration@0.0.19": {
         "files": [
             {
                 "relativePath": "src/find-configuration/find-configuration.ts",
@@ -508,7 +508,7 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "bit.test-envs/babel-dependencies@0.0.18": {
+    "bit.test-envs/babel-dependencies@0.0.19": {
         "files": [
             {
                 "relativePath": "src/babel-dependencies/babel-dependencies.ts",
@@ -531,5 +531,5 @@
         "origin": "AUTHORED",
         "exported": true
     },
-    "version": "13.0.5-dev.10"
+    "version": "13.0.5-dev.11"
 }
